refactor : month,year store 관리
