+++ index.js
@@ -1,11 +1,11 @@
 // js/index.js
 import { initDate } from "./components/header/dateRender.js";
-import { initRenderDummyData } from "./components/main/ledger/entries/renderDummyData.js";
+import { initRenderDummyData } from "./pages/main/ledger/entries/renderDummyData.js";
 import { initInputFormListener } from "./components/inputForm/inputFormItems/totalListener.js";
-import { deleteEntry } from "./components/main/ledger/entries/entry.js";
-import { initFilterButtons } from "./components/main/ledger/entries/filterEntries.js";
-import { initViewSwitcher } from "./components/main/viewSwitcher.js";
-import { updateCalendarTotalAmount } from "./components/main/calendar/calendarTotalAmount.js";
+import { deleteEntry } from "./pages/main/ledger/entries/entry.js";
+import { initFilterButtons } from "./pages/main/ledger/entries/filterEntries.js";
+import { initViewSwitcher } from "./pages/main/viewSwitcher.js";
+import { updateCalendarTotalAmount } from "./pages/main/calendar/calendarTotalAmount.js";
 import createHeader from "./components/header/header.js";
 import createInputForm from "./components/inputForm/input-form.js";
