+++ pages/main/viewSwitcher.js
